🚚 move brain external api secrets
