--- backend/repository/external_api_secret/__init__.py
@@ -1,2 +1,0 @@
-from .create_secret import create_secret
-from .read_secret import read_secret
